Depth Sorting Seems to be Working

+++ IsometricGame/source/IsometricCoord.cs
@@ -13,13 +13,13 @@ namespace IsometricGame.Source.AIsometricCoord
         {
             float x = (aPosition.X - aPosition.Y) * (TileWidth / 2);
             float y = (aPosition.X + aPosition.Y) * (TileHeight / 2);
-            return new Vector2(x, y);
+            return new Vector2((int)x, (int)y);
         }
         public static Vector2 IsoToMap(Vector2 aPosition)
         {
             float x = ((aPosition.X / (TileWidth / 2)) + (aPosition.Y / (TileHeight / 2))) / 2; //Valores provisorios para testes
             float y = ((aPosition.Y / (TileHeight / 2)) - (aPosition.X / (TileWidth / 2))) / 2;
-            return new Vector2(x, y);
+            return new Vector2((int)x, (int)y);
         }
     }
 }
